REFACTOR: Crinado o TrackStatus para fazer o mapeamento dos trechos da pista
